Read on for details!
Created a Custom String Scanner to deal with the shitty type of input
that the course guys have provided for testing the code. The all new
SetScanner passes with flying colours. To test it right now, copy past
this and press enter when your run the code: Dirk = {470423, 376329282,
235188141, 9447047, 94,  0, 141, 188  }

Happy Coding!